Compare two dependent groups

compare_dependent_groups provides numerous descriptive estimators and bootstrap confidence intervals for comparison between paired pre and post. compare_dependent_groups function is a "wrapper" for bmbstats function.

compare_dependent_groups(
  pre,
  post,
  SESOI_lower = SESOI_lower_dependent_func,
  SESOI_upper = SESOI_upper_dependent_func,
  estimator_function = dependent_groups_estimators,
  control = model_control(),
  na.rm = FALSE
)

Arguments

pre
Numeric vector

post
Numeric vector

SESOI_lower
Lower smallest effect size of interest threshold

SESOI_upper
Upper smallest effect size of interest threshold

estimator_function
Function that takes pre, post, and na.rm parameters and return named numeric vector with parameter estimates. Default is dependent_groups_estimators. User can write their own function with needed estimators for which bootstrap confidence intervals are needed

control
Control object returned from model_control function. Use boot_type, boot_samples, boot_strata to setup bootstrap.

na.rm
Should NAs be removed? Default is FALSE

Examples

data("bench_press_data")
pre <- bench_press_data$`Pre-test`
post <- bench_press_data$`Post-test`
compare_dependent_groups(pre, post, SESOI_lower = -5, SESOI_upper = 5)

#> Bootstraping: 2000 resamples

#> Calculating confidence intervals...

#> Warning: boot::boot.ci returned error or NULL when estimating CIs for SESOI lower estimator. Returning NAs for upper and lower CIs

#> [1] "All values of t are equal to  5 \n Cannot calculate confidence intervals"

#> Warning: boot::boot.ci returned error or NULL when estimating CIs for SESOI upper estimator. Returning NAs for upper and lower CIs

#> [1] "All values of t are equal to  10 \n Cannot calculate confidence intervals"

#> Warning: boot::boot.ci returned error or NULL when estimating CIs for SESOI range estimator. Returning NAs for upper and lower CIs

#> Done!

#> Bootstrap with 2000 resamples and 95% bca confidence intervals.
#> 
#>             estimator       value       lower       upper
#>           SESOI lower  -5.0000000          NA          NA
#>           SESOI upper   5.0000000          NA          NA
#>           SESOI range  10.0000000          NA          NA
#>           Mean change   5.5341489  1.83957169   8.8748919
#>             SD change   8.0501496  6.58011756  10.1943995
#>            %CV change 145.4631905 81.70400697 557.9322807
#>              % change   5.7533085  1.92727212   9.2106754
#>                 Ratio   1.0575331  1.01927272   1.0921068
#>             Cohen's d   0.6359037  0.18406313   1.0825663
#>                  CLES   0.6527598  0.54208345   0.7526957
#>                   OVL   0.7505214  0.58858566   0.9268186
#>  Mean change to SESOI   0.5534149  0.18395717   0.8874892
#>    SD change to SESOI   0.8050150  0.65801176   1.0194399
#>                pLower   0.1031451  0.03293031   0.2210039
#>           pEquivalent   0.3707501  0.26343315   0.4833775
#>               pHigher   0.5261049  0.33814603   0.7042765