Compare two independent groups

compare_independent_groups provides numerous descriptive estimators and bootstrap confidence intervals for comparison between group_a and group_b. compare_independent_groups function is a "wrapper" for bmbstats function.

compare_independent_groups(
  group_a,
  group_b,
  SESOI_lower = SESOI_lower_independent_func,
  SESOI_upper = SESOI_upper_independent_func,
  estimator_function = independent_groups_estimators,
  control = model_control(),
  na.rm = FALSE
)

Arguments

group_a
Numeric vector. This group represents baseline/control, observed variable, Pre-test in the paired design, or "practical" measure

group_b
Numeric vector. This group represents experimental, predicted variable, Post-test in the paired design, or "criterion" measure

SESOI_lower
Lower smallest effect size of interest threshold

SESOI_upper
Upper smallest effect size of interest threshold

estimator_function
Function that takes group_a, group_b, and na.rm parameters and return named numeric vector with parameter estimates. Default is independent_groups_estimators. User can write their own function with needed estimators for which bootstrap confidence intervals are needed

control
Control object returned from model_control function. Use boot_type, boot_samples, boot_strata to setup bootstrap.

na.rm
Should NAs be removed? Default is FALSE

Examples

data("height_data")
group_a <- height_data$Height[height_data$Gender == "Female"]
group_b <- height_data$Height[height_data$Gender == "Male"]
compare_independent_groups(group_a, group_b, SESOI_lower = -2.5, SESOI_upper = 2.5)

#> Bootstraping: 2000 resamples

#> Calculating confidence intervals...

#> Warning: boot::boot.ci returned error or NULL when estimating CIs for SESOI lower estimator. Returning NAs for upper and lower CIs

#> [1] "All values of t are equal to  2.5 \n Cannot calculate confidence intervals"

#> Warning: boot::boot.ci returned error or NULL when estimating CIs for SESOI upper estimator. Returning NAs for upper and lower CIs

#> [1] "All values of t are equal to  5 \n Cannot calculate confidence intervals"

#> Warning: boot::boot.ci returned error or NULL when estimating CIs for SESOI range estimator. Returning NAs for upper and lower CIs

#> Done!

#> Bootstrap with 2000 resamples and 95% bca confidence intervals.
#> 
#>           estimator       value       lower       upper
#>         SESOI lower -2.50000000          NA          NA
#>         SESOI upper  2.50000000          NA          NA
#>         SESOI range  5.00000000          NA          NA
#>           Mean diff 12.72521903  9.32103640  16.2507439
#>             SD diff 12.41402468 11.10890479  14.0886183
#>           SD pooled  8.77804103  7.85518191   9.9621576
#>            %CV diff 97.55450689 74.90802261 139.3775217
#>              % diff  7.79837084  5.67474889  10.0917848
#>               Ratio  1.07798371  1.05674749   1.1009178
#>           Cohen's d  1.44966502  1.01479494   1.8910657
#>                CLES  0.84733444  0.76347233   0.9092152
#>                 OVL  0.46855479  0.34462428   0.6118895
#>  Mean diff to SESOI  2.54504381  1.86420728   3.2501488
#>    SD diff to SESOI  2.48280494  2.22178096   2.8177237
#>              pLower  0.11148343  0.06291350   0.1853073
#>         pEquivalent  0.09457649  0.06713448   0.1217429
#>             pHigher  0.79394008  0.69988469   0.8684475