Describe relationship between two dependent groups

describe_relationship provides numerous descriptive estimators and bootstrap confidence intervals for relationship between paired predictor and outcome groups. describe_relationship function is a "wrapper" for bmbstats function.

describe_relationship(
  predictor,
  outcome,
  SESOI_lower = SESOI_lower_dependent_func,
  SESOI_upper = SESOI_upper_dependent_func,
  estimator_function = relationship_lm_estimators,
  control = model_control(),
  na.rm = FALSE
)

Arguments

predictor
Numeric vector

outcome
Vector

SESOI_lower
Lower smallest effect size of interest threshold

SESOI_upper
Upper smallest effect size of interest threshold

estimator_function
Function that takes predictor, outcome, and na.rm parameters and return named numeric vector with parameter estimates. Default is relationship_lm_estimators. User can write their own function with needed estimators for which bootstrap confidence intervals are needed

control
Control object returned from model_control function. Use boot_type, boot_samples, boot_strata to setup bootstrap.

na.rm
Should NAs be removed? Default is FALSE

Examples

data("yoyo_mas_data")
predictor <- yoyo_mas_data$YoYoIR1
outcome <- yoyo_mas_data$MAS

describe_relationship(predictor, outcome, SESOI_lower = -0.5, SESOI_upper = 0.5)

#> Bootstraping: 2000 resamples

#> Calculating confidence intervals...

#> Warning: boot::boot.ci returned error or NULL when estimating CIs for SESOI lower estimator. Returning NAs for upper and lower CIs

#> [1] "All values of t are equal to  0.5 \n Cannot calculate confidence intervals"

#> Warning: boot::boot.ci returned error or NULL when estimating CIs for SESOI upper estimator. Returning NAs for upper and lower CIs

#> [1] "All values of t are equal to  1 \n Cannot calculate confidence intervals"

#> Warning: boot::boot.ci returned error or NULL when estimating CIs for SESOI range estimator. Returning NAs for upper and lower CIs

#> Done!

#> Bootstrap with 2000 resamples and 95% bca confidence intervals.
#> 
#>     estimator        value        lower        upper
#>   SESOI lower -0.500000000           NA           NA
#>   SESOI upper  0.500000000           NA           NA
#>   SESOI range  1.000000000           NA           NA
#>     Intercept 13.159948349 12.795885062 13.533904164
#>         Slope  0.001402145  0.001096886  0.001678131
#>           RSE  0.218737182  0.181876509  0.270622184
#>   Pearson's r  0.857917432  0.712651772  0.927915833
#>     R Squared  0.736022320  0.507947744  0.861037856
#>  SESOI to RSE  4.571696467  3.695266244  5.501598338
#>          PPER  0.970243050  0.925107239  0.989794130